% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_with_gnomad_freqs.r
\name{annotate_with_gnomad_freqs}
\alias{annotate_with_gnomad_freqs}
\title{Annotate a dataset with GnomAD frequencies}
\usage{
annotate_with_gnomad_freqs(dataset)
}
\arguments{
\item{dataset}{dataset to annotate with GnomAD freqs; must have either chr, pos, ref, alt or txconsq}
}
\value{
Returns a data.table
}
\description{
Annotate a dataset with GnomAD frequencies
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
