% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIN_test_WES.r
\name{BIN_test_WES}
\alias{BIN_test_WES}
\title{Calculate a p-values from the BIN-test across the whole-exome}
\usage{
BIN_test_WES(
  cases,
  controls,
  case_coverage = NULL,
  control_coverage = NULL,
  cov_threshold = 0.5,
  covstats = F,
  messages = T
)
}
\arguments{
\item{cases}{case data in format: data.table(aff, symbol, protein_position, ac)}

\item{controls}{control data in format: data.table(aff, symbol, protein_position, ac)}

\item{case_coverage}{optional coverage data for cases in format: data.table(symbol, protein_position, over_10)}

\item{control_coverage}{optional coverage data for controls in format: data.table(symbol, protein_position, over_10)}

\item{cov_threshold}{threshold at which to exclude a residue position from the analysis (choose 0 to keep all residues)}

\item{covstats}{include additional coverage information as a column in results?}

\item{messages}{print messages to the terminal (e.g. how many variants removed by coverage)}
}
\value{
Returns a data.table: symbol, p-value, covstats (optional)
}
\description{
Calculate a p-values from the BIN-test across the whole-exome
}
\details{
This method splits each gene into k = n ^ (2/5) bins where n is the total number of observed variants in the gene. To control for uneven coverage cell counts are adjusted by the reciprocal of the mean 10X coverage across the bin.
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
