% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden_test.r
\name{burden_test}
\alias{burden_test}
\title{Carry out a Fisher's exact test}
\usage{
burden_test(
  n1 = NULL,
  n2 = NULL,
  ss1 = NULL,
  ss2 = NULL,
  pval = T,
  case_coverage = NULL,
  control_coverage = NULL,
  cov_threshold = 0.5,
  alternative = "greater"
)
}
\arguments{
\item{n1}{number of carriers cases}

\item{n2}{number of carriers controls}

\item{ss1}{sample size cases}

\item{ss2}{sample size controls}

\item{pval}{return p-value or full output?}

\item{case_coverage}{optional coverage data for cases in format: data.table(symbol, protein_position, over_10)}

\item{control_coverage}{optional coverage data for controls in format: data.table(symbol, protein_position, over_10)}

\item{cov_threshold}{threshold at which to exclude a residue position from the analysis (choose 0 to keep all residues)}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less".}
}
\value{
Returns an object of class htest - use $p for p-value
}
\description{
Burden test with formatting step to make a contingency table
}
\details{
Fisher.test from base R with extra formatting steps.

For a one-sided test i.e. excess in cases, use the default alternative="greater"

if pval=F
Use $p for p-value
Use $est for odds-ratio
Use $conf.int[1:2] for 95% CI
}
\examples{
# For the simple case where you have the vector status (binary case or control) and a vector of whether they
have a variant 'car' (presumably after filtering) then;

N = 100
car = sample(c(0, 1), N, rep=T)
status = sample(c(0, 1), N, rep=T)

burden_test(car, status)

burden_test(n1=10, n2=20, ss1=100, ss2=100)
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
