% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burden_test_WES.r
\name{burden_test_WES}
\alias{burden_test_WES}
\title{Calculate Fisher's-exact p-values across the whole-exome}
\usage{
burden_test_WES(
  cases,
  controls,
  cases_ss = NULL,
  controls_ss = NULL,
  case_coverage = NULL,
  control_coverage = NULL,
  cov_threshold = 0.5,
  alternative = "greater",
  covstats = F,
  messages = T
)
}
\arguments{
\item{cases}{case data in format: data.table(aff, symbol, protein_position, ac)}

\item{controls}{control data in format: data.table(aff, symbol, protein_position, ac)}

\item{cases_ss}{sample sizes for cases either a scalar for all genes or a data.table with two columns: symbol, ss}

\item{controls_ss}{sample sizes for controls either a scalar for all genes or a data.table with two columns: symbol, ss}

\item{case_coverage}{optional coverage data for cases in format: data.table(symbol, protein_position, over_10)}

\item{control_coverage}{optional coverage data for controls in format: data.table(symbol, protein_position, over_10)}

\item{cov_threshold}{threshold at which to exclude a residue position from the analysis (choose 0 to keep all residues)}

\item{alternative}{seearch for excess variant counts in cases only? (or "two-sided")}

\item{covstats}{include additional coverage information as a column in results?}

\item{messages}{print messages to the terminal (e.g. how many variants removed by coverage)}
}
\value{
Returns a data.table: symbol, p-value, covstats (optional)
}
\description{
Calculate Fisher's-exact p-values across the whole-exome
}
\details{
Calculates a Fisher's-exact test with coverage control across an exome set. Yates continuity correction used for 0 counts. Residue positions with mean 10X coverage less than "cov_threshold" excluded from the analysis.
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
