% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canon_txs-data.r
\docType{data}
\name{canon_txs}
\alias{canon_txs}
\title{Canonical transcripts used in this package}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 19856 rows and 2 columns.
}
\usage{
data(canon_txs)
}
\description{
Canonical transcripts used in this package
}
\keyword{datasets}
