% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinvar_cases-data.r
\docType{data}
\name{clinvar_cases}
\alias{clinvar_cases}
\title{ClinVAR variants}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 214061 rows and 8 columns.
}
\usage{
data(clinvar_cases)
}
\description{
ClinVAR variants
}
\keyword{datasets}
