% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterBurden_WES.r
\name{ClusterBurden_WES}
\alias{ClusterBurden_WES}
\title{ClusterBurden test for burden and position of rare-missense variants}
\usage{
ClusterBurden_WES(
  cases,
  controls,
  case_coverage = NULL,
  control_coverage = NULL,
  cases_ss = NULL,
  controls_ss = NULL,
  cov_threshold = 0.5,
  alternative = "greater",
  covstats = F,
  messages = T
)
}
\arguments{
\item{cases}{case data in format: data.table(aff, symbol, protein_position, ac)}

\item{controls}{control data in format: data.table(aff, symbol, protein_position, ac)}

\item{case_coverage}{optional coverage data for cases in format: data.table(symbol, protein_position, over_10)}

\item{control_coverage}{optional coverage data for controls in format: data.table(symbol, protein_position, over_10)}

\item{cases_ss}{sample sizes for cases either a scalar for all genes or a data.table with two columns: symbol, ss}

\item{controls_ss}{sample sizes for controls either a scalar for all genes or a data.table with two columns: symbol, ss}

\item{cov_threshold}{threshold at which to exclude a residue position from the analysis (choose 0 to keep all residues)}

\item{covstats}{include additional coverage information as a column in results?}

\item{messages}{print messages to the terminal (e.g. how many variants removed by coverage)}
}
\value{
A data.table with columns: symbol, ncases, ncontrols, case_ss, control_ss, binp, burdp, clusterburden
}
\description{
ClusterBurden test for burden and position of rare-missense variants
}
\details{
Perform ClusterBurden test across an exome set
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
