% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vector-data.r
\docType{data}
\name{color_vector}
\alias{color_vector}
\title{Personal colour palette}
\format{
An object of class \code{character} of length 92.
}
\usage{
data(color_vector)
}
\description{
Personal colour palette
}
\keyword{datasets}
