% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ps.r
\name{combine_ps}
\alias{combine_ps}
\title{Combine p-values}
\usage{
combine_ps(burden_pval, position_pval, method = "chisq")
}
\arguments{
\item{burden_pval}{P-value from a burden test such as Fisher's exact-test}

\item{position_pval}{P-value from a position test such as Chi-squared two-sample test}

\item{method}{Method to combine p-values. "chisq" for Fisher's method. "z-score" for Stouffer's method.}
}
\value{
Returns a single unnamed p-value of type num
}
\description{
Combine p-values using chi-squared or z-score combination method.
}
\details{
This is a method to combine p-values using either Fisher's method or Stouffer's method.

Usage is simple however you must first calculate burden and position p-values.
Burden p-values can be calculated using ClusterBurden::burden_test() or any other RVAT that does not use positional information
Position p-values can be calculated using ClusterBurden::BIN_test() or any other position test that is not sensitive to burden

If any p-values are zero then the function returns zero.

The functions have been adapted from the R package metap but re-written here to reduce dependencies.
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
