% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econtrols-data.r
\docType{data}
\name{econtrols}
\alias{econtrols}
\title{Control group - GnomAD exomes v2}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 5360408 rows and 12 columns.
}
\usage{
data(econtrols)
}
\description{
Control group - GnomAD exomes v2
}
\keyword{datasets}
