% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econtrols_nonpass-data.r
\docType{data}
\name{econtrols_nonpass}
\alias{econtrols_nonpass}
\title{GnomAD exomes v2 variants that did not pass filters - for frequency annotation only}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 600933 rows and 9 columns.
}
\usage{
data(econtrols_nonpass)
}
\description{
GnomAD exomes v2 variants that did not pass filters - for frequency annotation only
}
\keyword{datasets}
