% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exome_cov-data.r
\docType{data}
\name{exome_cov}
\alias{exome_cov}
\title{Coverage file for GnomAD exomes at gene-level}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11119849 rows and 3 columns.
}
\usage{
data(exome_cov)
}
\description{
Coverage file for GnomAD exomes at gene-level
}
\keyword{datasets}
