% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_genes.r
\name{find_genes}
\alias{find_genes}
\title{Select a subset of genes to use in an WES analysis}
\usage{
find_genes(n_genes, clndn_regex)
}
\arguments{
\item{n_genes}{how many random null genes?}

\item{clndn_regex}{regular expression to find disease genes from ClinVAR}
}
\description{
Select a subset of genes to use in an WES analysis
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
