% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRanges.r
\name{findRanges}
\alias{findRanges}
\title{Find contiguous ranges of values in a vector of integers}
\usage{
findRanges(x, output = "dt")
}
\arguments{
\item{x}{a vector of integers}

\item{output}{whether to output a data.table with columns start and end ("dt") or string ("str")}
}
\description{
Find contiguous ranges of values in a vector of integers
}
\details{
if output = "dt" then a data.table with the start and end positions of contiguous ranges of integers in the vector.
if output = "str" then each range is presented in a single string separated by commas e.g. "1-3,6-10,17-19"
}
\examples{
x = sample(1:10, 5)
findRanges(x, "dt")
findRanges(x, "str")
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
