% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_coverage.r
\name{format_coverage}
\alias{format_coverage}
\title{Format coverage files from genome positions (chrom, pos) to residue positions (symbol, index)}
\usage{
format_coverage(coverage)
}
\arguments{
\item{coverage}{a coverage file with 3 columnms: chrom, pos, tenX (proportion of samples with 10X coverage)}
}
\description{
Format coverage files from genome positions (chrom, pos) to residue positions (symbol, index)
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
