% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcontrols-data.r
\docType{data}
\name{gcontrols}
\alias{gcontrols}
\title{Control group - GnomAD genomes v2}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1444119 rows and 13 columns.
}
\usage{
data(gcontrols)
}
\description{
Control group - GnomAD genomes v2
}
\keyword{datasets}
