% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcontrols_nonpass-data.r
\docType{data}
\name{gcontrols_nonpass}
\alias{gcontrols_nonpass}
\title{GnomAD genomes v2 variants that did not pass filters - for frequency annotation only}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 125575 rows and 10 columns.
}
\usage{
data(gcontrols_nonpass)
}
\description{
GnomAD genomes v2 variants that did not pass filters - for frequency annotation only
}
\keyword{datasets}
