% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_locations-data.r
\docType{data}
\name{gene_locations}
\alias{gene_locations}
\title{Genomic coordinates of all canonical genes}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 19856 rows and 3 columns.
}
\usage{
data(gene_locations)
}
\description{
Genomic coordinates of all canonical genes
}
\keyword{datasets}
