% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_cov-data.r
\docType{data}
\name{genome_cov}
\alias{genome_cov}
\title{Coverage file for GnomAD genomes at a gene-level}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11098650 rows and 3 columns.
}
\usage{
data(genome_cov)
}
\description{
Coverage file for GnomAD genomes at a gene-level
}
\keyword{datasets}
