% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grmap_reduced-data.r
\docType{data}
\name{grmap_reduced}
\alias{grmap_reduced}
\title{Map between genome position and protein position in canonical genes}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 33359547 rows and 4 columns.
}
\usage{
data(grmap_reduced)
}
\description{
Map between genome position and protein position in canonical genes
}
\keyword{datasets}
