% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.r
\name{manhattan}
\alias{manhattan}
\title{A Manhattan plot for p-values derived from a ClusterBurden WES analysis}
\usage{
manhattan(pvals, test, n_genes, SCALE = 1)
}
\arguments{
\item{pvals}{pvals data.table from ClusterBurden WES analysis}

\item{test}{either bin-test, burden or clusterburden (not case sensitive)}
}
\value{
Returns a ggplot object
}
\description{
A Manhattan plot for p-values derived from a ClusterBurden WES analysis
}
\details{
A Manhattan plot for p-values resulting from a ClusterBurden analysis
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
