% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nresidues-data.r
\docType{data}
\name{nresidues}
\alias{nresidues}
\title{Length (in residues) of all canonical proteins}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 21675 rows and 2 columns.
}
\usage{
data(nresidues)
}
\description{
Length (in residues) of all canonical proteins
}
\keyword{datasets}
