% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_features.r
\name{plot_features}
\alias{plot_features}
\title{Plot sequence features such as domains alongside distributions of case and control variant residues}
\usage{
plot_features(
  genename,
  pvals,
  cases,
  controls,
  case_coverage,
  control_coverage,
  cov_threshold = 0.5,
  features = c("Compositional-bias", "Coiled-coil", "Repeat", "Zinc-Finger", "Region",
    "Motif", "Domain", "Topological domain", "Intramembrane", "Transmembrane",
    "Cross-link", "Disulphide", "Glycosylation", "Lipidation", "Peptide", "Propeptide",
    "Signal peptide", "Transit peptide", "Structure"),
  SCALE = 1
)
}
\arguments{
\item{genename}{gene to plot}

\item{pvals}{pvals data.table from ClusterBurden WES analysis}

\item{cases}{case data in format: data.table(aff, symbol, protein_position, ac)}

\item{controls}{control data in format: data.table(aff, symbol, protein_position, ac)}

\item{case_coverage}{optional coverage data for cases in format: data.table(symbol, protein_position, over_10)}

\item{control_coverage}{optional coverage data for controls in format: data.table(symbol, protein_position, over_10)}

\item{cov_threshold}{threshold at which to exclude a residue position from the analysis (choose 0 to keep all residues)}

\item{features}{which sequence features to plot e.g. c("Domain", "Region", "Structure")}

\item{SCALE}{scale of the plot}
}
\value{
Returns a ggplot object
}
\description{
Plot sequence features such as domains alongside distributions of case and control variant residues
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
