% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.r
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot for chi-squared residuals}
\usage{
plot_residuals(chisq)
}
\arguments{
\item{chisq}{output from chi-squared two-sample test.}
}
\value{
Returns an object of class ggplot.
}
\description{
This function allows you to plot the residuals of a chi-squared two-sample test to infer bins
that differ significantly from expected values.
}
\details{
This method plots residuals as a 2 x m table where m is the number of bins in the chisq.test.

Residuals are standardized by the standard deviation of the residuals.
Within each cell of the table is a circle whose color and size relate to the size of the residual.
Residuals that are significant at a nominal 0.05 (two-tailed) after Bonferonni correction are highlighted with an asterix.
}
\examples{
nresidues = 1000 # length of the protein
probs = rexp(nresidues)^2
case_probs = probs * rep(c(1, 3, 1), c(200, 200, 600))
control_probs = probs * rep(c(2, 1, 2), c(200, 200, 600))

case_residues = sample(1:nresidues, 100, rep=T, case_probs)
control_residues = sample(1:nresidues, 100, rep=T, control_probs)

chisq = BIN_test(case_residues, control_residues)
plot_residuals(chisq)
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
