% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signif_distribs.r
\name{plot_signif_distribs}
\alias{plot_signif_distribs}
\title{Plot the distributions of case and control variant residues across multiple genes}
\usage{
plot_signif_distribs(
  pvals,
  cases,
  controls,
  case_coverage,
  control_coverage,
  test = "BIN-test",
  cov_threshold = 0.5,
  n_genes = 20,
  SCALE = 1
)
}
\arguments{
\item{pvals}{pvals data.table from ClusterBurden WES analysis with at least symbol and p-value column for test of interest}

\item{cases}{case data in format: data.table(aff, symbol, protein_position, ac)}

\item{controls}{control data in format: data.table(aff, symbol, protein_position, ac)}

\item{case_coverage}{optional coverage data for cases in format: data.table(symbol, protein_position, over_10)}

\item{control_coverage}{optional coverage data for controls in format: data.table(symbol, protein_position, over_10)}

\item{test}{either bin-test, burden or clusterburden (not case sensitive)}

\item{cov_threshold}{threshold at which to exclude a residue position from the analysis (choose 0 to keep all residues)}

\item{n_genes}{number of genes to plot e.g. n_genes = 5 plots the 5 most significant genes}

\item{SCALE}{scale of the plot}
}
\value{
Returns a ggplot object
}
\description{
Plot the distributions of case and control variant residues across multiple genes
}
\details{
For each gene (n_genes=?): produces a stripchart using geom_jitter for both cases and controls as well as a gaussian density line for each cohort. If coverage files are provided then these are included as a rug underneath and above the case and control variants and are
colored by their coverage level (< cov_threshold, < 80% 10X, < 90% 10X, > 90% 10X)
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
