% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_uncovered_residues.r
\name{remove_uncovered_residues}
\alias{remove_uncovered_residues}
\title{Exclude residue positions from a dataset based on coverage files}
\usage{
remove_uncovered_residues(
  dataset,
  case_coverage,
  control_coverage,
  cov_threshold,
  messages
)
}
\arguments{
\item{dataset}{case-control data.table with at least columns: symbol and protein_position}

\item{case_coverage}{optional coverage data for cases in format: data.table(symbol, index, over_10)}

\item{control_coverage}{optional coverage data for controls in format: data.table(symbol, index, over_10)}

\item{cov_threshold}{threshold at which to exclude a residue position from the analysis (choose 0 to keep all residues)}

\item{messages}{print messages to the terminal (e.g. how many variants removed by coverage)}
}
\description{
Exclude residue positions from a dataset based on coverage files
}
\details{
Variant residues are removed from both cases and controls if they have low coverage in either case or control coverage. These positions are also removed from the coverage files to make downstream esimation of regional coverage accurate in light of variants removed.
}
\author{
Adam Waring - adam.waring@msdtc.ox.ac.uk
}
\keyword{RVAT,}
\keyword{case-control,}
\keyword{cluster,}
\keyword{distribution,}
\keyword{gene}
\keyword{genetics,}
