% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_features-data.r
\docType{data}
\name{seq_features}
\alias{seq_features}
\title{Sequences features for all proteins downloaded from Uniprot}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 306918 rows and 5 columns.
}
\usage{
data(seq_features)
}
\description{
Sequences features for all proteins downloaded from Uniprot
}
\keyword{datasets}
