% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x10_gnomad-data.r
\docType{data}
\name{x10_gnomad}
\alias{x10_gnomad}
\title{Proportion of samples with at least x10 for genes in GnomAD exomes and genomes}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 19831 rows and 3 columns.
}
\usage{
data(x10_gnomad)
}
\description{
Proportion of samples with at least x10 for genes in GnomAD exomes and genomes
}
\keyword{datasets}
